Refactored according to Maeve and Victor's comments. Bumped version for beta release.
