+++ app/models/g5_authenticatable/user.rb
@@ -37,8 +37,12 @@ module G5Authenticatable
       end
     end
 
+    def selectable_clients
+      G5Updatable::SelectableClientPolicy::Scope.new(self, G5Updatable::Client).resolve
+    end
+
     def clients
-      G5Updatable::ClientPolicy::Scope.new(self, G5Updatable::Client).clients_from_client_and_location_roles
+      G5Updatable::ClientPolicy::Scope.new(self, G5Updatable::Client).resolve
     end
 
     def locations
