+++ app/policies/g5_updatable/location_policy.rb
@@ -1,11 +1,14 @@
+# The policy will resolve to all locations that a user has access. This includes locations that a user has roles for
+# but also those locations for which a user has a client role.  Global roles grant access to ALL locations.
 module G5Updatable
   class LocationPolicy < G5Authenticatable::BasePolicy
     class Scope < G5Authenticatable::BasePolicy::BaseScope
 
       def resolve
-        return scope.all if has_global_role?
-        scope.where(id: location_roles.map(&:resource_id))
+        locations_from_client_location_roles
       end
+
+      private
 
       def location_roles
         G5Authenticatable::Role
@@ -15,19 +18,22 @@ module G5Updatable
                  G5Updatable::Location.name, user.id)
       end
 
-      def clients_from_location_roles
-        G5Updatable::Client
-          .joins('INNER JOIN g5_updatable_locations as l on l.client_uid=g5_updatable_clients.uid')
-          .joins('INNER JOIN g5_authenticatable_roles as r on l.id=r.resource_id')
-          .joins('INNer JOIN g5_authenticatable_users_roles as ur on r.id=ur.role_id')
-          .where('r.resource_type = ? and ur.user_id = ?',
-                 G5Updatable::Location.name, user.id)
-          .group('g5_updatable_clients.id')
+      def locations_from_client_roles
+        G5Updatable::Location
+          .joins('INNER JOIN g5_updatable_clients as c on g5_updatable_locations.client_uid=c.uid')
+          .joins('INNER JOIN g5_authenticatable_roles as r on r.resource_id=c.id')
+          .joins('INNER JOIN g5_authenticatable_users_roles as ur on r.id=ur.role_id')
+          .where('ur.user_id=?',user.id)
+          .where('r.resource_type=?', G5Updatable::Client.name)
       end
 
-      def has_global_role?
-        G5Authenticatable::BasePolicy.new(user).has_global_role?
+
+      def locations_from_client_location_roles
+        return scope.all if has_global_role?
+        location_ids = locations_from_client_roles.map(&:id) | location_roles.map(&:resource_id)
+        G5Updatable::Location.where(id: location_ids)
       end
+
     end
 
   end
