+++ lib/g5_authenticatable/version.rb
@@ -1,3 +1,3 @@
 module G5Authenticatable
-  VERSION = '0.7.5'
+  VERSION = '0.8.0.beta1'
 end
