+++ spec/policies/client_policy_spec.rb
@@ -58,67 +58,4 @@ describe G5Updatable::ClientPolicy do
       end
     end
   end
-
-  describe '.clients_from_client_and_location_roles', focus: true do
-    subject { G5Updatable::ClientPolicy::Scope.new(user, G5Updatable::Client).clients_from_client_and_location_roles }
-
-    let!(:location_1) { FactoryGirl.create(:g5_updatable_location, client: client_1) }
-    let!(:location_2) { FactoryGirl.create(:g5_updatable_location, client: client_1) }
-
-    let!(:location_3) { FactoryGirl.create(:g5_updatable_location, client: client_2) }
-    let!(:location_4) { FactoryGirl.create(:g5_updatable_location, client: client_2) }
-
-    context 'with global role' do
-      before { user.add_role :admin }
-      it 'returns all clients' do
-        expect(subject.length).to eq(3)
-        expect(subject).to include(client_1)
-        expect(subject).to include(client_2)
-        expect(subject).to include(client_3)
-      end
-    end
-
-    context 'with role for location and for client that location belongs to' do
-      before do
-        user.add_role :admin, location_1
-        user.add_role :admin, client_1
-      end
-      it 'returns 1 client' do
-        expect(subject.length).to eq(1)
-        expect(subject).to include(client_1)
-      end
-    end
-
-    context 'with role for location and client that location does not belong to' do
-      before do
-        user.add_role :admin, location_1
-        user.add_role :admin, client_2
-      end
-      it 'returns 1 client' do
-        expect(subject.length).to eq(2)
-        expect(subject).to include(client_1)
-        expect(subject).to include(client_2)
-      end
-    end
-
-    context 'with a client role' do
-      before do
-        user.add_role :admin, client_2
-      end
-      it 'returns 1 client' do
-        expect(subject.length).to eq(1)
-        expect(subject).to include(client_2)
-      end
-    end
-
-    context 'with a location role' do
-      before do
-        user.add_role :admin, location_1
-      end
-      it 'returns 1 client' do
-        expect(subject.length).to eq(1)
-        expect(subject).to include(client_1)
-      end
-    end
-  end
 end
